feat(new class): add hero class
This commit add the hero class in a new module and within two methods. The constructor method and the fight method.
In the constructor method it adds the name, starting health (default to 100) and the current health.
Inside the fight method the istructions are to choose one winner randomnly, for now.