feat(new class): adds a weapon class and the method to add it to the hero's abilities
This commit add a new class called Weapon which is set as a subclass of Ability
and overwrite the attack method from Ability. In the Hero class a method called add_weapon
is created to add the weapond to the list of abilities that are going to be used
when the hero attacks.